+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-# master
+# v0.12.1 to-be-released
 
 ## Fixed
 
 * `Constructor#try` rescues `ArgumentError` (raised in cases like `Integer('foo')`) (flash-gordon)
+* `#constructor` works correctly for default and enum types (solnic)
 
 [Compare v0.12.0...master](https://github.com/dry-rb/dry-types/compare/v0.12.0...master)